full updATE ADMIN AND ROOM
hti

+++ room-service/room-service/src/main/java/com/example/room_service/Service/RoomService.java
@@ -20,18 +20,28 @@ public class RoomService {
         return roomRepository.findAll();
     }
 
+    public  Room getRoomById(int id) {
+        Optional<Room> Room = roomRepository.findById(id);
+        if (Room.isPresent()) {
+            return Room.get();
+        }
+        return null;
+    }
+
     public Room createRoom(Room room) {
         return roomRepository.save(room);
     }
 
-    public Optional<Room> getRoomById(int id) {
-        return roomRepository.findById(id);
+    public Room updateRoom(Room room) {
+        return roomRepository.save(room)    ;
     }
 
-    public void deleteRoom(int id) {
-        roomRepository.deleteById(id);
+    public Room deleteRoomById(int id) {
+        Optional<Room> Room = roomRepository.findById(id);
+        if (Room.isPresent()) {
+            roomRepository.deleteById(id);
+        }
+        return Room.get();
     }
-
-
 
 }
